Separate out settings components
